+++ index.html
@@ -21,7 +21,6 @@
     <div class="main-container">
         <main class="menu-container">
             <section class="filters-container" id="filters">
-                <button class="filter">filter</button>
             </section>
             <div class="products-container" id="products">
                 <div class="product-container">
@@ -59,7 +58,7 @@
             </div>
             <!-- Recibo (así se debería ver) -->
             <div id="receipt-container">
-                <button class="close-button" id="close-receipt"><img src="./assets/img/close.svg" alt="close"></button>
+                <button class="close-button" id=" close-receipt"><img src="./assets/img/close.svg" alt="close"></button>
                 <h2 class="receipt-title">Recibo</h2>
                 <div class="receipt-product" id="receipt-product">
                     <h3>Título</h3>
